svelte: resize for xterm
Allows the terminal to auto adjust when the window size changes. This is
useful when a user wants to, for example, copy a large portion of text
from the buffer, therefore they will zoom out a lot.

Previously they would have had to refresh the page for the terminal to
fill a large portion of their window, but in this example, the buffer
would be cleared, ultimately defeating the purpose of doing so.